Refactor Notices to reusable hook
